+++ src/ui/options/Options.tsx
@@ -1,3 +1,4 @@
+import Preferences from '@/common/services/preferences';
 import ChromeLocalStorage from '@/storage/chrome/chrome-local-storage';
 import ChromeSyncStorage from '@/storage/chrome/chrome-sync-storage';
 import React, { useEffect, useState } from 'react';
@@ -6,7 +7,6 @@ import * as styles from './Options.module.css';
 import classNames from 'classnames';
 import * as psl from 'psl';
 import { ParsedDomain } from 'psl';
-import Preferences from '@/preferences';
 
 Preferences.initDefaults(new ChromeSyncStorage(), new ChromeLocalStorage());
